+++ e2e/test-env/src/client/RicosTestApp/editorPlugins.ts
@@ -95,7 +95,6 @@ const defaultConfigs: TestAppConfig['pluginsConfig'] = {
   audio: {
     handleFileSelection: audioHandlers.handleFileSelection,
     fetchData: mockFetchUrlPreviewData(),
-    enableCustomUploadOnMobile: true,
     getAudioUrl: src => `https://static.wixstatic.com/${src.id}`,
     exposeButtons: [
       audioButtonsTypes.audio,